refactor clients and extensions (#691)
