--- packages/stream_video/lib/src/coordinator/coordinator_ws.dart
@@ -1,13 +1,0 @@
-import '../shared_emitter.dart';
-import '../ws/ws.dart';
-import 'models/coordinator_events.dart';
-
-abstract class CoordinatorWebSocket extends StreamWebSocket {
-  CoordinatorWebSocket(
-    super.url, {
-    super.protocols,
-    super.tag,
-  });
-
-  SharedEmitter<CoordinatorEvent> get events;
-}
