+++ packages/stream_video/lib/src/lifecycle/lifecycle_state.dart
@@ -7,9 +7,7 @@ enum LifecycleState {
 
   @override
   String toString() => name;
-}
 
-extension LifecycleStateX on LifecycleState {
   bool get isResumed => this == LifecycleState.resumed;
 
   bool get isPaused => this == LifecycleState.paused;
