+++ packages/stream_video/lib/src/models/call_participant_state.dart
@@ -1,9 +1,11 @@
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
+import 'user_info.dart';
 
 import '../sfu/data/models/sfu_connection_quality.dart';
 import '../sfu/data/models/sfu_track_type.dart';
 import '../sorting/call_participant_sorting_presets.dart';
+import '../utils/string.dart';
 import 'call_reaction.dart';
 import 'call_track_state.dart';
 import 'viewport_visibility.dart';
@@ -166,4 +168,11 @@ class CallParticipantState
   bool get isScreenShareEnabled {
     return !(screenShareTrack?.muted ?? true);
   }
+
+  UserInfo toUserInfo() => UserInfo(
+    id: userId,
+    role: roles.firstOrNull ?? '',
+    name: name.ifEmpty(() => userId),
+    image: image,
+  );
 }
