+++ packages/stream_video_flutter/lib/src/utils/extensions.dart
@@ -49,42 +49,3 @@ extension DoubleExtension on double {
     return max(minimumValue, min(this, maximumValue));
   }
 }
-
-extension CallParticipantExtension on CallParticipantState {
-  /// Maps [CallParticipantState] into [UserInfo].
-  UserInfo toUserInfo() => UserInfo(
-        id: userId,
-        role: roles.firstOrNull ?? '',
-        name: name.ifEmpty(() => userId),
-        image: image,
-      );
-}
-
-extension CallStatusExtension on CallStatus {
-  /// Returns human readable text representation of [CallStatus].
-  String toStatusString() {
-    if (isIdle) {
-      return 'Idle';
-    } else if (isOutgoing) {
-      return 'Outgoing';
-    } else if (isIncoming) {
-      return 'Incoming';
-    } else if (isJoining) {
-      return 'Joining';
-    } else if (isJoined) {
-      return 'Joined';
-    } else if (isReconnecting) {
-      return 'Reconnecting';
-    } else if (isMigrating) {
-      return 'Migrating';
-    } else if (isConnecting) {
-      return 'Connecting';
-    } else if (isConnected) {
-      return 'Connected';
-    } else if (isDisconnected) {
-      return 'Disconnected';
-    } else {
-      return '';
-    }
-  }
-}
